requirements.txt
azure-cosmos
Faker
